// src/lib/basicsearch/types.ts
export type BasicProduct = {
  __typename?: 'IProduct';
  id: string;
  name?: string;
  description?: string;
  price?: number;
  regular_price?: number;
  discount?: number;
  image_url?: string;
  thumbnail_url?: string;
  categories?: string[];
  brands?: string[];
  colors?: string[];
  materials?: string[];
  parent_categories?: string[];
  typologies?: string[];
  units?: string[];
  uses?: string[];
  stock?: number;
  total_sales?: number;
  is_on_sale?: boolean;
  created_at?: string;
  updated_at?: string;
  external_id?: number;
  slug?: string;
  sku?: string;
  preciomts2?: number;
  sale_price?: number;
};